Chisel changes and other small tweaks

+++ src/main/java/net/tier1234/better_deco/block/custom/CustomBarrelBlock.java
@@ -6,9 +6,9 @@ import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.stats.Stats;
 import net.minecraft.util.RandomSource;
+import net.minecraft.world.Container;
 import net.minecraft.world.Containers;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.MenuProvider;
 import net.minecraft.world.entity.monster.piglin.PiglinAi;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
@@ -23,25 +23,26 @@ import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraft.world.level.block.state.properties.BooleanProperty;
 import net.minecraft.world.level.block.state.properties.DirectionProperty;
-import net.minecraft.world.level.block.state.properties.Property;
 import net.minecraft.world.phys.BlockHitResult;
 import net.tier1234.better_deco.block.entity.CustomBarrelBlockEntity;
 import net.tier1234.better_deco.block.entity.ModBlockEntities;
 
 public class CustomBarrelBlock extends BaseEntityBlock {
-    public static final MapCodec<net.minecraft.world.level.block.BarrelBlock> CODEC = simpleCodec(net.minecraft.world.level.block.BarrelBlock::new);
-    public static final DirectionProperty FACING;
-    public static final BooleanProperty OPEN;
+    public static final MapCodec<CustomBarrelBlock> CODEC = simpleCodec(CustomBarrelBlock::new);
+    public static final DirectionProperty FACING = BlockStateProperties.FACING;
+    public static final BooleanProperty OPEN = BlockStateProperties.OPEN;
 
-    public MapCodec<net.minecraft.world.level.block.BarrelBlock> codec() {
+    @Override
+    protected MapCodec<? extends BaseEntityBlock> codec() {
         return CODEC;
     }
 
     public CustomBarrelBlock(BlockBehaviour.Properties properties) {
         super(properties);
-        this.registerDefaultState((BlockState)((BlockState)((BlockState)this.stateDefinition.any()).setValue(FACING, Direction.NORTH)).setValue(OPEN, false));
+        this.registerDefaultState(this.stateDefinition.any()
+                .setValue(FACING, Direction.NORTH)
+                .setValue(OPEN, false));
     }
-
 
     @Override
     protected InteractionResult useWithoutItem(BlockState state, Level level, BlockPos pos,
@@ -50,8 +51,8 @@ public class CustomBarrelBlock extends BaseEntityBlock {
             return InteractionResult.SUCCESS;
         } else {
             BlockEntity blockEntity = level.getBlockEntity(pos);
-            if (blockEntity instanceof CustomBarrelBlockEntity) {
-                player.openMenu((CustomBarrelBlockEntity) blockEntity);
+            if (blockEntity instanceof CustomBarrelBlockEntity barrel) {
+                player.openMenu(barrel);
                 player.awardStat(Stats.OPEN_BARREL);
                 PiglinAi.angerNearbyPiglins(player, true);
                 return InteractionResult.CONSUME;
@@ -60,18 +61,23 @@ public class CustomBarrelBlock extends BaseEntityBlock {
         }
     }
 
-
-    protected void onRemove(BlockState state, Level level, BlockPos pos, BlockState newState, boolean isMoving) {
-        Containers.dropContentsOnDestroy(state, newState, level, pos);
-        super.onRemove(state, level, pos, newState, isMoving);
+    @Override
+    public void onRemove(BlockState state, Level level, BlockPos pos, BlockState newState, boolean isMoving) {
+        if (!state.is(newState.getBlock())) {
+            BlockEntity be = level.getBlockEntity(pos);
+            if (be instanceof Container container) {
+                Containers.dropContents(level, pos, container);
+                level.updateNeighbourForOutputSignal(pos, this);
+            }
+            super.onRemove(state, level, pos, newState, isMoving);
+        }
     }
 
-    protected void tick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random) {
-        BlockEntity blockentity = level.getBlockEntity(pos);
-        if (blockentity instanceof BarrelBlockEntity) {
-            ((BarrelBlockEntity)blockentity).recheckOpen();
+    @Override
+    public void tick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random) {
+        if (level.getBlockEntity(pos) instanceof BarrelBlockEntity barrel) {
+            barrel.recheckOpen();
         }
-
     }
 
     @Override
@@ -79,36 +85,40 @@ public class CustomBarrelBlock extends BaseEntityBlock {
         return ModBlockEntities.CUSTOM_BARREL_BE.get().create(pos, state);
     }
 
+    @Override
     protected RenderShape getRenderShape(BlockState state) {
         return RenderShape.MODEL;
     }
 
+    @Override
     protected boolean hasAnalogOutputSignal(BlockState state) {
         return true;
     }
 
-    protected int getAnalogOutputSignal(BlockState blockState, Level level, BlockPos pos) {
+    @Override
+    protected int getAnalogOutputSignal(BlockState state, Level level, BlockPos pos) {
         return AbstractContainerMenu.getRedstoneSignalFromBlockEntity(level.getBlockEntity(pos));
     }
 
+    @Override
     protected BlockState rotate(BlockState state, Rotation rotation) {
-        return (BlockState)state.setValue(FACING, rotation.rotate((Direction)state.getValue(FACING)));
+        return state.setValue(FACING, rotation.rotate(state.getValue(FACING)));
     }
 
+    @Override
     protected BlockState mirror(BlockState state, Mirror mirror) {
-        return state.rotate(mirror.getRotation((Direction)state.getValue(FACING)));
+        return state.rotate(mirror.getRotation(state.getValue(FACING)));
     }
 
+    @Override
     protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
-        builder.add(new Property[]{FACING, OPEN});
+        builder.add(FACING, OPEN);
     }
 
+    @Override
     public BlockState getStateForPlacement(BlockPlaceContext context) {
-        return (BlockState)this.defaultBlockState().setValue(FACING, context.getNearestLookingDirection().getOpposite());
-    }
-
-    static {
-        FACING = BlockStateProperties.FACING;
-        OPEN = BlockStateProperties.OPEN;
+        return this.defaultBlockState()
+                .setValue(FACING, context.getNearestLookingDirection().getOpposite())
+                .setValue(OPEN, false);
     }
 }
